+++ src/main/resources/db/changelog/changeset/user_V001__initial.sql
@@ -197,20 +197,3 @@ CREATE TABLE rating (
     CONSTRAINT fk_rater_id FOREIGN KEY (user_id) REFERENCES users (id),
     CONSTRAINT fk_event_rated_id FOREIGN KEY (event_id) REFERENCES event (id)
 );
-
-CREATE TABLE Transactions (
-    id BIGINT GENERATED BY DEFAULT AS IDENTITY PRIMARY KEY,
-    payment_number BIGINT NOT NULL UNIQUE,
-    amount DECIMAL(19, 2) NOT NULL,
-    message VARCHAR(255),
-    currency VARCHAR(3) NOT NULL,
-    verification_code INTEGER,
-    user_id BIGINT NOT NULL,
-    creating_date TIMESTAMP,
-    updated_at TIMESTAMP,
-    type VARCHAR(64) NOT NULL,
-    product_code VARCHAR(64) NOT NULL,
-    transaction_status VARCHAR(32) NOT NULL,
-
-    CONSTRAINT fk_transactions_user FOREIGN KEY (user_id) REFERENCES users (id)
-);
