+++ src/main/resources/db/changelog/changeset/user_V009_add_transaction_table.sql
@@ -1,0 +1,17 @@
+
+CREATE TABLE IF NOT EXISTS Transactions (
+    id BIGINT GENERATED BY DEFAULT AS IDENTITY PRIMARY KEY,
+    payment_number BIGINT NOT NULL UNIQUE,
+    amount DECIMAL(19, 2) NOT NULL,
+    message VARCHAR(255),
+    currency VARCHAR(3) NOT NULL,
+    verification_code INTEGER,
+    user_id BIGINT NOT NULL,
+    creating_date TIMESTAMP,
+    updated_at TIMESTAMP,
+    type VARCHAR(64) NOT NULL,
+    product_code VARCHAR(64) NOT NULL,
+    transaction_status VARCHAR(32) NOT NULL,
+
+    CONSTRAINT fk_transactions_user FOREIGN KEY (user_id) REFERENCES users (id)
+);
