+++ src/main/resources/db/changelog/db.changelog-master.yaml
@@ -15,3 +15,5 @@ databaseChangeLog:
       file: db/changelog/changeset/user_V007__user_premium.sql
   - include:
       file: db/changelog/changeset/user_V008__add_personal_info_tables.sql
+  - include:
+      file: db/changelog/changeset/user_V009_add_transaction_table.sql
