+++ src/test/java/school/faang/user_service/service/SkillServiceTest.java
@@ -19,6 +19,7 @@ import school.faang.user_service.repository.SkillRepository;
 import school.faang.user_service.repository.UserRepository;
 import school.faang.user_service.repository.UserSkillGuaranteeRepository;
 import school.faang.user_service.repository.recommendation.SkillOfferRepository;
+import school.faang.user_service.service.skill.SkillService;
 
 import java.util.List;
 import java.util.Optional;
